[MAINT]: Project relationships
Treat the relationship Project<->SeismicCatalog and
Project<->Injectionwell as real one to many relationships. Remove orphans
automatically.
As @heilukas mentioned, a project reflects an entity in the sense of a
*real* project such as e.g. COSEISMIQ.

Alternatively, a project might be considered as an application level
project. Then of course, it would make sense to treat the relationships
mentioned above as many to many.