[MAINT]: referencepoint and srs
Prepare the datamodel for a local coordinate system. Thus, introduce
referencepoint and a spatialreference system (srs) fields configured at
a project level. srs is represented as a string. It SHOULD be configured
in Proj4 notation. (Depending on the application code, alternatively WKT
might be used.)

Input data MUST be in WGS84 (srid=4326).