[ENH][BUG] Snapshots used for serializing
A bug was found in the ramsis code where if multiple model workers use
the same code and this code has to be serialized where alterations are
made (for instance the coordinate transformation from local to wgs84),
then The same object will be serialized. This leads to multiple
transformations happening on the same data which is not good. Now
snapshots are made of the data which makes a clone recursively. A few
alterations had to be made to allow snapshots of well plans to be made
and for all the data to be copied accurately. Also if a detached
instance does not have an attribute, this is handled in a try except
case so that deteched errors are caught.